+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/parsers/ms_merge.sql
@@ -10,3 +10,11 @@ WHEN NOT MATCHED THEN
 INSERT (UnitMeasureCode, Name)
 VALUES (src.UnitMeasureCode, src.Name)
 OUTPUT deleted.*, $action, inserted.* INTO #MyTempTable;
+
+MERGE l
+    USING ((SELECT @P, @C, @S) AS T (P, C, S)
+            JOIN P ON T.P = P.ID
+            JOIN C ON T.C = C.ID)
+    ON MATCH (P-(l)->C)
+WHEN NOT MATCHED THEN INSERT DEFAULT VALUES
+WHEN MATCHED THEN UPDATE SET S = @S;