+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added support for PostgreSQL 16.
+
 ### Changed
+
+- Improved parser rules for MS SQL.
 
 ### Fixed